// src/api/back/models/RegistrationDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    RegistrationPlayerDto,
    RegistrationPlayerDtoFromJSON,
    RegistrationPlayerDtoFromJSONTyped,
    RegistrationPlayerDtoToJSON,
} from './';

/**
 * 
 * @export
 * @interface RegistrationDto
 */
export interface RegistrationDto {
    /**
     * 
     * @type {number}
     * @memberof RegistrationDto
     */
    id: number;
    /**
     * 
     * @type {Array<RegistrationPlayerDto>}
     * @memberof RegistrationDto
     */
    players: Array<RegistrationPlayerDto>;
    /**
     * 
     * @type {string}
     * @memberof RegistrationDto
     */
    state: RegistrationDtoStateEnum;
    /**
     * 
     * @type {string}
     * @memberof RegistrationDto
     */
    title: string;
}

export function RegistrationDtoFromJSON(json: any): RegistrationDto {
    return RegistrationDtoFromJSONTyped(json, false);
}

export function RegistrationDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): RegistrationDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'players': ((json['players'] as Array<any>).map(RegistrationPlayerDtoFromJSON)),
        'state': json['state'],
        'title': json['title'],
    };
}

export function RegistrationDtoToJSON(value?: RegistrationDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'players': ((value.players as Array<any>).map(RegistrationPlayerDtoToJSON)),
        'state': value.state,
        'title': value.title,
    };
}

/**
* @export
* @enum {string}
*/
export enum RegistrationDtoStateEnum {
    CREATED = 'CREATED',
    PENDINGCONFIRMATION = 'PENDING_CONFIRMATION',
    CONFIRMED = 'CONFIRMED',
    DECLINED = 'DECLINED',
    TIMEDOUT = 'TIMED_OUT'
}